pyproject.toml
[tool.poetry]
name = "globoticket"
version = "0.1.0"
description = "Demo project for the Pluralsight course 'Building a REST API with Python 3'"
authors = ["James Santos-Calacat <jamessantosc@gmail.com>","Reindert-Jan Ekker <info@codesensei.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.1"}
sqlalchemy = "^2.0.20"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
